requirements.txt
# web/api
fastapi>=0.111,<1
uvicorn[standard]>=0.29
httpx>=0.27

# ingest & parsing
feedparser>=6.0
trafilatura>=1.6
beautifulsoup4>=4.12
lxml>=5.2
readability-lxml>=0.8.1
tldextract>=5.1

# nlp basics
spacy>=3.7,<3.8
langdetect>=1.0.9
dateparser>=1.2
python-dateutil>=2.9

# embeddings + retrieval
sentence-transformers>=3.0.1
rank-bm25>=0.2
chromadb>=0.5,<0.6

# optional: if you want local rerank later, use CrossEncoder from sentence-transformers (no extra pip)

# graph + clustering
networkx>=3.3
python-louvain>=0.16
scikit-learn>=1.4,<1.6


# data utils
pandas>=2.2
numpy<2

# config/logging/retries
pydantic>=2.8
pydantic-settings>=2.3
python-dotenv>=1.0
tenacity>=8.3
rich>=13.7

# gemini client (direct integration)
google-generativeai>=0.7
joblib>=1.4

jinja2>=3.1

